% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_heatmap.R
\name{resample_heatmap}
\alias{resample_heatmap}
\title{Resampled Kernel Density Heatmap}
\usage{
resample_heatmap(
  data,
  lat_col,
  lon_col,
  n_samples = NULL,
  col_data = NULL,
  filter_by = NULL,
  basemap = NULL,
  input_crs = 4326
)
}
\arguments{
\item{data}{A data frame with your point data. Needs to have latitude and longitude columns.}

\item{lat_col}{A string for the name of your latitude column}

\item{lon_col}{A string for the name of your longitude column}

\item{n_samples}{Optional. Integer to re-sample data from the \code{filter_by} column to a new size. Must be greater than the current number of values for \code{filter_by}.}

\item{col_data}{Optional. Name of the column you are using to filter your data.}

\item{filter_by}{Optional. Name of the value to filter the \code{col_data} column by values.}

\item{basemap}{Optional. Spatial base-map used for plotting.}

\item{input_crs}{Optional. Specify the coordinate reference system used in your input data. Default coordinate system is EPSG:4326 (WGS 84).}
}
\value{
A ggplot2 heatmap — you can print it, save it, or add more layers if you want.
}
\description{
Create a 2D Kernel Density Heatmap by Resampling Small Datasets with Optional Filter and Basemap

This function is an alternative to the make_heatmap() function.
This function will resample data to create a heatmap when there isn't sufficient data.
Can be used to resample filtered data or the whole data set.
When the resample size is not specified, it will multiply the number of rows by 1000 as a default.
It generates a smooth heatmap from latitude and longitude coordinates.You can also add a base-map.
}
\examples{
#Resample all data without specifying `n_samples`.
resample_heatmap(data = my_data, lat_col = "latitude", lon_col = "longitude")

#Resample filtered data to 10000 and plot heatmap with resamlped data.
resample_heatmap(data = my_data, lat_col = "latitude", lon_col = "longitude", col_data = "offense", filter_by = "Homicide", n_samples = 10000)

#Resample all data to 100000 and plot heatmap with resampled data.
resample_heatmap(data = my_data, lat_col = "latitude", lon_col = "longitude", n_samples = 10000)

}
