% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_kriging.R
\name{density_kriging}
\alias{density_kriging}
\title{Kriging-Based Density Map}
\usage{
density_kriging(
  data,
  lat_col,
  lon_col,
  input_crs = 4326,
  col_data = NULL,
  filter_by = NULL,
  grid_size = 500,
  variogram_model = "Sph"
)
}
\arguments{
\item{data}{A data frame with your point data. Needs to have latitude and longitude columns.}

\item{lat_col}{A string for the name of your latitude column}

\item{lon_col}{A string for the name of your longitude column}

\item{input_crs}{Optional. Specify the coordinate reference system used in your input data. Default coordinate system is EPSG:4326 (WGS 84).}

\item{col_data}{Optional. Name of the column you are using to filter your data.}

\item{filter_by}{Optional. Name of the value to filter the \code{col_data} column by values.}

\item{grid_size}{Optional. The size in m^2 of your grids for counting/density. If left blank, it defaults to 500}

\item{variogram_model}{Optional. A specific vgm type fit for your  data's variogram (e.g., "Gau"). If left blank, it defaults to a spherical model}

\item{basemap}{Optional. Spatial base-map used for plotting.}
}
\value{
A base R density plot with predicted densities of points for each grid — you can print it, save it, or add more layers if you want.
}
\description{
Create a map of predicted density through kriging

This function generates a predicted density map of a categorical value from latitude and longitude coordinates and a list of observations.
It can be used to generate kriging results for categorical data that would otherwise be unable to be analyzed.
Using this function lets you infer data about missing points to better round out your dataset
You can adjust this data to fit the CRS, grid size for analysis, and variogram type to your specific needs.
}
\examples{
# Plot density in standard grids and variogram types
density_kriging(my_data, lon_col = "longitude", lat_col = "latitude", input_crs = 4326)

 # Plot density at a different grid size or variagram style - for example 250 m^2 grid and a gaussian variogram
density_kriging(my_data, lon_col = "longitude", lat_col = "latitude", input_crs = 4326,
 grid_size = 250, variogram_model = "Gau")

 # Plot density with a just one kind of data -for example, only "Theft"
density_kriging(my_data, lon_col = "longitude", lat_col = "latitude", input_crs = 4326,
 col_data = "offense", filter_by = "Theft")
}
