% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Kernel Density Heatmap}
\usage{
make_heatmap(
  data,
  lat_col,
  lon_col,
  col_data = NULL,
  filter_by = NULL,
  basemap = NULL,
  input_crs = 4326
)
}
\arguments{
\item{data}{Required. A data frame with your point data. Needs to have latitude and longitude columns.}

\item{lat_col}{Required. A string for the name of your latitude column.}

\item{lon_col}{Required. A string for the name of your longitude column.}

\item{col_data}{Optional. Name of the column you are using to filter your data. Can be categorical or factor data for filtering purposes.}

\item{filter_by}{Optional. Name of the value to filter the \code{col_data} column by values. Only works when \code{col_data} is also included.}

\item{basemap}{Optional.  A spatial object (e.g., an \code{sf} or \code{raster} object) to use as background for plot.}

\item{input_crs}{Optional. Specify a different EPSG code or PROJ string for CRS used in your input data. Default coordinate system is EPSG:4326 (WGS 84).}
}
\value{
A ggplot2 heatmap — you can print it, save it, or add more layers if you want.
}
\description{
Create a 2D Kernel Density Heatmap from Latitude and Longitude with Optional Filter and Basemap

This function generates a smooth heatmap from latitude and longitude coordinates.
You can use it to map any kind of point data (like crimes, wildlife sightings, etc.).
If you want, you can filter the data to show just one type (like only "Robbery" or just "Bird A").
You can also include a background map (basemap), but that part is optional.
}
\examples{
# Sample Data:
my_data <- read.csv(system.file("extdata", "dc_sample.csv", package = "geoSweepR"))
my_basemap <- terra::rast(system.file("extdata", "my_basemap.tif", package = "geoSweepR"))

# Plot all data without filtering.
make_heatmap(my_data, lat_col = "latitude", lon_col = "longitude")

# Plot filtered data
# — Example uses only "Theft".
make_heatmap(my_data, lat_col = "latitude", lon_col = "longitude",
             col_data = "offense", filter_by = "Theft")

# Plot and add a basemap layer underneath.
make_heatmap(my_data, lat_col = "latitude", lon_col = "longitude",
             basemap = my_basemap)
}
